pyproject.toml
[project]
name = "cfupdater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.14",
    "onepassword-sdk>=0.2.1",
]
